requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.1.6
django-channels==0.7.0
django-ckeditor==6.7.2
django-js-asset==3.0.1
idna==3.10
kombu==5.5.4
oauthlib==3.2.2
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
